test(project): use docker for visual tests
