+++ packages/project-nitro-twig/tests/backstop/bitmaps_reference/nitro_test_Homepage_0_m-example_1_tablet.png
