+++ packages/project-nitro-twig/tests/backstop/bitmaps_reference/nitro_test_Homepage_1_m-example__n2_0_phone.png
