+++ packages/project-nitro/gulpfile.js
@@ -15,6 +15,3 @@ gulp.task('develop', ['watch-assets', 'watch-serve']);
 gulp.task('dump-views', getTask('dump-views'));
 gulp.task('lint-accessibility', ['dump-views'], getTask('lint-accessibility'));
 gulp.task('lint-html', ['dump-views'], getTask('lint-html'));
-gulp.task('visual-approve', getTask('visual-approve'));
-gulp.task('visual-reference', getTask('visual-reference'));
-gulp.task('visual-test', getTask('visual-test'));
