+++ packages/project-nitro/package.json
@@ -35,9 +35,11 @@
     "test:cypress": "npm run build && cross-env PORT=8888 npm-run-all --parallel --race test:cypress:*",
     "test:cypress:serve": "npm run prod:serve",
     "test:cypress:test": "cypress run --project ./tests/cypress/",
-    "visual-approve": "gulp visual-approve",
-    "visual-reference": "npm run build && gulp visual-reference",
-    "visual-test": "npm run build && gulp visual-test",
+    "visual-approve": "backstop approve --config=tests/backstop/backstop.config.js --docker",
+    "visual-reference": "npm run build && backstop reference --config=tests/backstop/backstop.config.js --docker",
+    "visual-test": "npm run build && cross-env PORT=8889 npm-run-all --parallel --race visual-test:*",
+    "visual-test:serve": "npm run prod:serve",
+    "visual-test:test": "backstop test --config=tests/backstop/backstop.config.js --docker",
     "\n# BUILD/RELEASE ------- ": "",
     "build": "npm-run-all build:*",
     "build:clean": "rimraf public/assets",
@@ -95,6 +97,7 @@
     "@nitro/exporter": "*",
     "@nitro/gulp": "*",
     "@nitro/webpack": "*",
+    "backstopjs": "4.0.6",
     "commitizen": "3.1.1",
     "config": "3.1.0",
     "cross-env": "5.2.0",
